add environmental_functions.py, rest functions moved here.
