+++ climsight.py
@@ -48,6 +48,12 @@ from climate_functions import (
    load_data,
    extract_climate_data
 )
+from environmental_functions import (
+   fetch_biodiversity,
+   load_nat_haz_data,
+   filter_events_within_square,
+   plot_disaster_counts
+)
 
 with open('config.yml', 'r') as file:
     config = yaml.safe_load(file)
@@ -113,210 +119,7 @@ class StreamHandler(BaseCallbackHandler):
         else:
             raise ValueError(f"Invalid display_method: {self.display_method}")
 
-@st.cache_data
-def fetch_biodiversity(lon, lat):
-    """
-    Fetches biodiversity data for a given longitude and latitude using the GBIF API.
-
-    Args:
-    - lon (float): The longitude of the location to fetch biodiversity data for.
-    - lat (float): The latitude of the location to fetch biodiversity data for.
-
-    Returns:
-    - data (dict): A dictionary containing the biodiversity data for the specified location.
-    """
-    gbif_api_url = "https://api.gbif.org/v1/occurrence/search"
-    params = {
-        "decimalLatitude": lat,
-        "decimalLongitude": lon,
-    }
-    response = requests.get(gbif_api_url, params=params, timeout=3)
-    biodiv = response.json()
-    biodiv_set = set()
-    if biodiv['results']:
-        for record in biodiv['results']:
-            if 'genericName' in record and record.get('taxonRank') != 'UNRANKED':
-                biodiv_set.add(record['genericName'])
-        biodiversity = ', '.join(list(biodiv_set))
-    else:
-        biodiversity = "Not known"
-    return biodiversity
-
-
-# @st.cache_data
-# def load_data():
-#     hist = xr.open_mfdataset(f"{data_path}/AWI_CM_mm_historical*.nc", compat="override")
-#     future = xr.open_mfdataset(f"{data_path}/AWI_CM_mm_ssp585*.nc", compat="override")
-#     return hist, future
-
-# def convert_to_mm_per_month(monthly_precip_kg_m2_s1):
-#     days_in_months = np.array([31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31])
-#     return monthly_precip_kg_m2_s1 * 60 * 60 * 24 * days_in_months
-
-
-# @st.cache_data
-# def extract_climate_data(lat, lon, _hist, _future):
-#     """
-#     Extracts climate data for a given latitude and longitude from historical and future datasets.
-
-#     Args:
-#     - lat (float): Latitude of the location to extract data for.
-#     - lon (float): Longitude of the location to extract data for.
-#     - _hist (xarray.Dataset): Historical climate dataset.
-#     - _future (xarray.Dataset): Future climate dataset.
-
-#     Returns:
-#     - df (pandas.DataFrame): DataFrame containing present day and future temperature, precipitation, and wind speed data for each month of the year.
-#     - data_dict (dict): Dictionary containing string representations of the extracted climate data.
-#     """
-#     hist_temp = hist.sel(lat=lat, lon=lon, method="nearest")["tas"].values - 273.15
-#     hist_temp_str = np.array2string(hist_temp.ravel(), precision=3, max_line_width=100)[
-#         1:-1
-#     ]
-
-#     hist_pr = hist.sel(lat=lat, lon=lon, method="nearest")["pr"].values
-#     hist_pr = convert_to_mm_per_month(hist_pr)
-
-#     hist_pr_str = np.array2string(hist_pr.ravel(), precision=3, max_line_width=100)[
-#         1:-1
-#     ]
-
-#     hist_uas = hist.sel(lat=lat, lon=lon, method="nearest")["uas"].values
-#     hist_uas_str = np.array2string(hist_uas.ravel(), precision=3, max_line_width=100)[
-#         1:-1
-#     ]
-
-#     hist_vas = hist.sel(lat=lat, lon=lon, method="nearest")["vas"].values
-#     hist_vas_str = np.array2string(hist_vas.ravel(), precision=3, max_line_width=100)[
-#         1:-1
-#     ]
-
-#     future_temp = future.sel(lat=lat, lon=lon, method="nearest")["tas"].values - 273.15
-#     future_temp_str = np.array2string(
-#         future_temp.ravel(), precision=3, max_line_width=100
-#     )[1:-1]
-
-#     future_pr = future.sel(lat=lat, lon=lon, method="nearest")["pr"].values
-#     future_pr = convert_to_mm_per_month(future_pr)
-#     future_pr_str = np.array2string(future_pr.ravel(), precision=3, max_line_width=100)[
-#         1:-1
-#     ]
-
-#     future_uas = future.sel(lat=lat, lon=lon, method="nearest")["uas"].values
-#     future_uas_str = np.array2string(
-#         future_uas.ravel(), precision=3, max_line_width=100
-#     )[1:-1]
-
-#     future_vas = future.sel(lat=lat, lon=lon, method="nearest")["vas"].values
-#     future_vas_str = np.array2string(
-#         future_vas.ravel(), precision=3, max_line_width=100
-#     )[1:-1]
-#     df = pd.DataFrame(
-#         {
-#             "Present Day Temperature": hist_temp[0, 0, :],
-#             "Future Temperature": future_temp[0, 0, :],
-#             "Present Day Precipitation": hist_pr[0, 0, :],
-#             "Future Precipitation": future_pr[0, 0, :],
-#             "Present Day Wind Speed": np.hypot(hist_uas[0, 0, :], hist_vas[0, 0, :]),
-#             "Future Wind Speed": np.hypot(future_uas[0, 0, :], future_vas[0, 0, :]),
-#             "Month": range(1, 13),
-#         }
-#     )
-#     data_dict = {
-#         "hist_temp": hist_temp_str,
-#         "hist_pr": hist_pr_str,
-#         "hist_uas": hist_uas_str,
-#         "hist_vas": hist_vas_str,
-#         "future_temp": future_temp_str,
-#         "future_pr": future_pr_str,
-#         "future_uas": future_uas_str,
-#         "future_vas": future_vas_str,
-#     }
-#     return df, data_dict
-
-
 hist, future = load_data(data_path)
-
-@st.cache_data
-def load_nat_haz_data(haz_path):
-    """
-    Load natural hazard data from a CSV file and filter relevant columns.
-
-    Args:
-    - haz_path (str): File path to the CSV file containing natural hazard data.
-
-    Returns:
-    - pandas.DataFrame: Dataset with selected columns ('country', 'year', 'geolocation', 'disastertype', 'latitude', 'longitude').
-    """
-
-    haz_dat = pd.read_csv(haz_path)
-
-    # reduce data set to only contain relevant columns
-    columns_to_keep = ['country', 'year', 'geolocation', 'disastertype', 'latitude', 'longitude']
-    haz_dat = haz_dat.loc[:, columns_to_keep]
-
-    return(haz_dat)
-
-@st.cache_data
-def filter_events_within_square(lat, lon, haz_path, distance_from_event):
-    """
-    Filter events within a square of given distance from the center point.
-
-    Args:
-    - lat (float): Latitude of the center point (rounded to 3 decimal places)
-    - lon (float): Longitude of the center point (rounded to 3 decimal places)
-    - haz_dat (pandas.DataFrame): Original dataset.
-    - distance_from_event (float): Distance in kilometers to form a square.
-
-    Returns:
-    - pandas.DataFrame: Reduced dataset containing only events within the square.
-    """
-
-    haz_dat = load_nat_haz_data(haz_path)
-
-    # Calculate the boundaries of the square
-    lat_min, lat_max = lat - (distance_from_event / 111), lat + (distance_from_event / 111)
-    lon_min, lon_max = lon - (distance_from_event / (111 * np.cos(np.radians(lat)))), lon + (distance_from_event / (111 * np.cos(np.radians(lat))))
-
-    # Filter events within the square
-    filtered_haz_dat = haz_dat[
-        (haz_dat['latitude'] >= lat_min) & (haz_dat['latitude'] <= lat_max) &
-        (haz_dat['longitude'] >= lon_min) & (haz_dat['longitude'] <= lon_max)
-    ]
-
-    prompt_haz_dat = filtered_haz_dat.drop(columns=['country', 'geolocation', 'latitude', 'longitude'])
-
-    return filtered_haz_dat, prompt_haz_dat
-
-@st.cache_data
-def plot_disaster_counts(filtered_events):
-    """
-    Plot the number of different disaster types over a time period for the selected location (within 5km radius).
-
-    Args:
-    - filtered_events: Only those natural hazard events that were within a 5 km (or whatever other value is set for distance_from_event) radius of the clicked location.
-    Returns:
-    - figure: bar plot with results
-    """
-    if not filtered_events.empty:
-        # Group by 'year' and 'disastertype' and count occurrences
-        disaster_counts = filtered_events.groupby(['year', 'disastertype']).size().unstack(fill_value=0)
-        place = filtered_events['geolocation'].unique()
-
-        # create figure and axes
-        fig, ax = plt.subplots(figsize=(10,6))
-        
-        # Plotting the bar chart
-        disaster_counts.plot(kind='bar', stacked=False, ax=ax, figsize=(10,6), colormap='viridis')
-        ax.yaxis.set_major_locator(MaxNLocator(integer=True))
-        plt.title('Count of different disaster types in ' + place[0] + ' over time')
-        plt.xlabel('Year')
-        plt.ylabel('Count')
-        plt.legend(title='Disaster Type')
-
-        return fig
-    else:
-        return None
 
 st.title(
     " :cyclone: \
